amazonSpider1.1/Spider/requirement.txt
backcall==0.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
ConcurrentLogHandler==0.9.1
decorator==4.2.1
demjson==2.2.4
gevent==1.2.2
greenlet==0.4.13
idna==2.6
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.11.1
lxml==4.2.1
numpy==1.14.2
opencv-python==3.4.0.12
parso==0.1.1
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
pytz==2018.4
Pygments==2.2.0
redis==2.10.6
requests==2.18.4
scipy==1.0.1
selenium==3.11.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
fake-useragent==0.1.10
retrying==1.3.3
